% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnR.R
\docType{package}
\name{AURAkNN-package}
\alias{AURAkNN}
\alias{AURAkNN-package}
\title{AURAkNN: Accelerated Ultra-rapid R Architecture for k-NN imputation}
\description{
AURAkNN implements high-performance k-Nearest Neighbors (kNN) imputation algorithm. It features a highly optimized C backend with OpenMP multithreading, supporting Euclidean, Manhattan, and Gower distances, and operates in two distinct modes: fast and precise.
}
\author{
\strong{Maintainer}: Michał Kalwas \email{michal.kalwas.stud@pw.edu.pl}

}
