% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnR.R
\name{kNN_impute}
\alias{kNN_impute}
\title{Fast and Precise k-Nearest Neighbours Imputation for Mixed Data}
\usage{
kNN_impute(
  data,
  k = 5,
  metric = c("gower", "euclidean", "manhattan"),
  mode = c("fast", "precise"),
  num_fun = c("median", "mean"),
  threads = NULL,
  maxColNa = 0.8,
  maxRowNa = 1
)
}
\arguments{
\item{data}{A data.frame or matrix, which contains a missing values (NAs) to be imputed.}

\item{k}{An integer specifying the number of neighbours. Default: 5.}

\item{metric}{A string specifying the  distance metric to use.
Allowed values are \code{"gower"} (default), \code{"euclidean"}, or \code{"manhattan"}.}

\item{mode}{A string specifying the search algorithm.
\code{"fast"} calculates global neighbors
\code{"precise"} recalculates neighbors for each missing cell. Uses weights in Gower distance metric. (default)}

\item{num_fun}{A string specifying the function used to impute missing cells.
Allowed values are \code{"median"} (default) or \code{"mean"}.}

\item{threads}{An integer specifying the number of threads used by the function.
Defaults to all available CPU cores minus one.}

\item{maxColNa}{A numeric value between 0.0 and 1.0. Columns with fraction of missing values greater
than this number will be removed. Default value is 0.8.}

\item{maxRowNa}{A numeric value between 0.0 and 1.0. Rows with fraction of missing values greater
than this number will be removed. Default value is 1.0.}
}
\value{
A data.frame with missing values replaced by imputed estimates.
}
\description{
Imputes missing values (NAs) in a given dataset using kNN algorithm.
The function supports Euclidean and Manhattan distances, designed for fast numeric imputation
and a Gower distance metric for mixed data. The function operates in two modes: "Fast", which is meant to be used
for rapid, memory-efficient imputation and "Precise mode", which maximizes the accuracy of the imputation.
}
\examples{
\dontrun{
 df <- data.frame(
   PatientId = as.integer(1:8),
   HeartRate = c(72, 85, 70, 88, NA, 60, 78, 86),
   Cholesterol = c(190, 240, 185, 250, 200, NA, 195, 255),
   RiskLevel = as.factor(c("Low", "High", "Low", "High", "Moderate", "Low", "Low", NA))
 )
 
 #Fast imputation using Euclidean distance
 kNNFast <- kNN_impute(df, k = 2, metric = "euclidean", mode = "fast")
 # kNNFast$Cholesterol[6]
 # [1] 192.5
 
 #Highly accurate imputation using Gower distance
 kNNPrecise <- kNN_impute(df, k = 2, metric = "gower", mode = "precise")
 # kNNPrecise$Risk_Level[8]
 # [1] High

}

}
